Completetion of HTML for Header and 'Main Content
This Commit includes the first version html of header and main content in
html only as per the specification document provided by TOP.

Note that this is only first draft and will likely need to be changed during CSS implementation.
This is last backup pre-CSS Implementation.

Placeholder Image also uploaded to be used on webpage. this will likely be used for the duration
of the project.