Redesign of TOP Landing Page to (imaginary) Dog Walking Service
This update bring a change in the direction for the design of the TOP Landing
page. This is no longer following the exact scope of the document and will instead
be following my own color schemes / images / text etc. to allow for a more personal approach.

With this said. Changes with this commit are as follows:

- Updated Logo created which was generated using Midjourney AI (Subscribed) and then converted to an SVG.
- Color Palette for website generated using https://coolors.co/ , see ./for info/Color Palette
- General update of website color scheme and content to suit above
- Implementation of further navbar links and :hover effect on each.

Current TODO: List / Bug list
TODO: HTML and CSS Secondary Content Wrapper (Some random information flexbox with images /text)
TODO: HTML and CSS Tertiary Content Wrapper (Quotes subsection)
TODO: HTML and CSS Sub-footer (Call to action)